Move macro to `lib.rs`, fix tests
